* refactored the code to separate the rescue functions from instance functions * moved some functions from utils.py to other modules that are more cohesive
