+++ gce_rescue/messages_test.py
@@ -16,7 +16,7 @@
 
 from absl.testing import absltest
 from gce_rescue import messages
-from gce_rescue.rescue import Instance
+from gce_rescue.gce import Instance
 from gce_rescue.test.mocks import mock_api_object, MOCK_TEST_VM
