+++ gce_rescue/tasks/metadata_test.py
@@ -15,9 +15,8 @@
 """Test code for metadata.py."""
 
 from absl.testing import absltest
-from absl import logging
 
-from gce_rescue.rescue import Instance
+from gce_rescue.gce import Instance
 from gce_rescue.tasks import metadata
 from gce_rescue.test.mocks import  mock_api_object, MOCK_TEST_VM
