+++ gce_rescue/tasks/operations_test.py
@@ -15,7 +15,7 @@
 """Test code for operations.py."""
 
 from absl.testing import absltest
-from gce_rescue.rescue import Instance
+from gce_rescue.gce import Instance
 from gce_rescue.tasks.operations import start_instance, stop_instance
 from gce_rescue.test.mocks import mock_api_object, MOCK_TEST_VM
